wip: impl Serialize for Tweet via round_trip! macro
the idea behind the macro is that i want the Deserialize impl to work
either with the original Twitter JSON, *or* with the serialization
output from the derived Serialize impl. Since i can't attempt an
automatic serialization directly into Tweet as-is (it would try to read
it as Twitter's JSON) i decided to make a copy of the struct definition
and provide a transparent From impl for easy conversion. combine this
with an enum labeled with `#[serde(untagged)]` and serde will do the
"try either of these" logic for me!

i had to leave out Place and User because they had complicated
serialization logic that i wanted to parcel out to other commits (and in
User's case, since it also contains a Tweet, i would have to apply
Serialize to both of them at the same time for it to "fully" work).